requirements.txt
discord.py>=2.4.0
openpyxl>=3.1.2
python-dotenv>=1.0.1